Improvement of test harness
Main goal: Separate testing of executable from the tests of the output.
This way if the executable segaults, we know that is the issue, and if
it runs, but the output is bad, we know that is the issue immediately.

Each test directory then has two shell scripts that accomplish these
goals (I prefer shell scripts for debugging rather than using the
arcane CTest methods of debugging with log file in obscure locations,
having to invoke CTest itself with the right arguments, etc.)

To test the actual running of the code,  petscdiff instead of diff to
enable ignore the floats

This resulted in some minor changes to the scripts themselves.