Goal: get a working Intel build and improve documentation
An Intel version has been compiled on NERSC and passes all of
the tests (config file has been added).  This uncovered the sensitivity
of 'zlim' to floating point errors.  A small shift in ylim was enough to
recover agreement, but a more robust method will probably be necessary
in other configurations.

INSTALL.rst is fairly complete now and RUN.rst has been started.
config_mac11.sh is a skelleton that needs to be filled out.