Goal: Build and test with MPI
A multiple time slice test is fully setup for parallel builds.  It
would probably be wise to add this option for all existing tests.
Summary data is only output for the first slice, so g, a, and m files
need to be compared for the rest (not yet automated).

The config scripts for Cori and Iris have been updated to build with
MPI.  Parallel netcdf was linked when available, but it not clear
whether it is used.

All MPI code is protected by compiler directives so some cleanup was in
order.

MPI versions continue to pass all debug checks and agree on the test
resutls.