Adding the cuda files
The cuda director corresponds to the "P-EFIT" files.  It contains many
hard-code values, is difficult to decipher, and is invoked via "option
8" in the interactive EFIT interface.

The fortran files efitdud129.f90 and getfnmdud.f90 are included in the
cuda directory.  If we really wanted to enable them, we would use the
mods from those files as well.  They also have hard-coded values.