Initializing variables and improving compiler performance.
All uninitiazed variables that are used in the tests have been set to
zero.  This eliminates all errors in the cray debug builds.

The CMAKE build types have been revised and now work for every
compiler.  This has prompted changing the config files for clusters
to build as release.

The install.rst now has a link to the google doc tutorial.  Some minor
additional changes were made to this and run.rst.

Further code cleanup was also performed.