Adding namelist reads from omas equilibrium files.
This is the first step in ongoing development.

This introduces a new operation mode (9) for efit which is only
available when HDF5 is linked.

This has only been tested on a single shot on a single machine.  Further
testing is required.

rbdry and zbdry are currently not read since they are not in database
files.  This will likely present issues for machine independence.

Next steps include:
  loading an initial solution profile from the omas files when available
  writing out results to omas files (overwriting?)
  using limiter and other data instead of ascii files?
  adding new tests/options for these capabilities