Replace input debug variables with pre-processor and add new warnings.
Run CMAKE with -DIDEBUG=# or -DJDEBUG="NAME" to achieve the outputs
that previously came from idebug and jdebug input variables.  This
reduces the number of if statements evaluated during runtime.  Warnings
are output if the input variables are set now that they are no longer
used (no change to namelists or files).

EFIT will now error out if a mesh size is requested which is
incompatible with Buneman's algorithm.  This limitation is avoided in
pefit so the warning will need to be updated when new matrix solvers are
implemented.

A new error condition was also added for MPI cases that have unused
processors.  A warning is output if the available processors are
unbalanced in terms of the number of time-slices they solve.

Note: these new errors were implemented to help explain cases which
would have otherwise failed in less obvious ways.  No functionality was
actually changed.