Making sphinx a CMake build target
This now builds the sphinx documentation in the build directory assuming
all of the dependencies are there.  With anaconda python, these are all
normal `conda install`.

Part of this moves `conf.py` into a CMake configure_file.  Although not
completely exploited, this allows one to do platform-specific
customizations to the sphinx build process.