Add tests and finish development of OMAS input
When hdf5 is linked every test will be run twice: first using the
standard input files (opt 2) and then using OMAS for input.

Intricacies of the different input variables and Fortran namelist
reads motivated a full rewrite of the input methods.  In order to
keep each variable read to a single line, new subroutines were
developed to more closely mimick the behavior of namelists.

The option to read inputs from OMAS was moved from option 9 to
option 1, since it was not being used.  This also aligns with the
primary uses for EFIT-AI.

The default length for q-profile targets (mqwant) was increased to
accommodate the variable sizes in NSTX input files.

A new section was added to the developer documentation outlining how
to add new tests.