Fixes for building with HDF5 on Iris.
The config scripts have been updated and minor cmake changes were
required.

Unfortunately the gnu version had to be rolled back in order to
use a working system HDF5 installation.  The gcc 9 config script
is retained in case admins add support for HDF5 with this compiler
or it is preferred for other purposes.

The intel and pgi compilers caught a few unallocated or uninitialized
variables which have now been fixed.