Code builds and runs but produces garbage
The code compiles and links without errors on GNU, Intel and PGI,
but GNU and Intel builds segfault at the first MDS+ calls, probably
because they are not supported by the system.

PGI builds can now run all of the snap input options.  When only a
single time-slice is choosen, the code will produce output files,
but all variables are zero (garbage).  When multiple time-slices are
selected, the code fails to find BCoil data for the later times.
All of these issues appear to stem from problems using the ptdata code
which is located external to the EFIT source, but is the key to
the DIIID MDS+ database.

Experts on the DIIID MDS+ data are likely needed to finish this
development.