Redo MPI Test command and add Git version
Rather than set an empty variable as the default MPI command used to
execute tests (also renamed), we set it as "mpirun" and reset it to be
empty if the build is serial.  As a result, cluster running slurm need
to set this in the config files (all updated).

Code to collect and write Git and build information to STDOUT has been
shamelessly pulled from NIMROD.  EFIT-AI now uses the git hashes to
uniquely denote versions instead of the build dates.  This seems like an
inconsequential departure from the legacy version in the name of better
provenance.