Keep efitversion as integer date
This is a legacy variable that is part of namelists reads and writes.
So trying to change it's type to align with git versioning caused
errors with PGI compilers.

The code now supports the date based nomenclature for this variable,
but in every other instance the git hash is used as a unique version
label.