Snap mode options work for efit01
Still working with Bill Meyer to get correct data from mselibs.

Some significant changes were required to make these options perform
properly.  Also fixed k-file issues that were/are present in the original
EFIT.  For some reason PGI builds with any compiler optimization above -O0
cause snap modes to seg fault so this has been set for all build options
until it can be sorted out.

Tests for the snap modes have been added and the incorrect DIIID k-files
that were used in other tests have been replaced (along with the results).
The kineticEFIT test still needs to be updated though.  Changes in the
convergence motivated a switch to different time slices for these tests.
Some changes (either to the software or Green's function tables on Iris)
have led to small numerical discrepancies in the results produced by the
system version of EFIT (the 20201123 version that we forked does not
produce the same results as it did when we forked, but EFIT-AI does...)
so we do not compare with those in the regression tests, but they are kept
around for manual comparisons.

Improved documentation when possible, including some of the discussion from
Slack.

Modernized the code in many places.