MSE snap options work
Many thanks to Bill Meyer for helping track down the issues with
FP types in our version of getdat.

All snap options appear to be fully functional for PGI compilers and
tests are setup.

The snap modes now produce identical results to the file modes.  This
required several fixes to problems that were in the original version
of EFIT.

Some code modernization and other improvements were made concurrently.