Updating the mhdin.dat for tests and making a support_file directory
The outdated mhdin.dat that was being used for testing has been
replaced with the correct version.  This brings all of the tests to
excellent agreement with the version of efit on Iris.  The test
results have been updated, but because of very small differences we
do not use the legacy efit results for the automated comparisons.

The share/data directory has been replaced with a share/support_file
directory which will contain all of the necessary machine info to run
efund and subsequently efit for any shot requested.  Nearly every
machine that has pre-computed Green function tables on the GA clusters
and NERSC is included.

The location that efit looks for fitweight.dat files has been changed
to better match where other machine files are located, since it does
not need to be shot dependent (can include info for multiple shot
ranges).