Need to use computed F-coil currents
Using the input F-coil currents reduces the number of iterations,
but does not capture a complete EFIT solution.  Unfortunately the
computed values are not written to the main part of a g-file or
included in the OMAS conversion in OMFIT (yet).  Instead we need to
read the out1 namelist from a g-file regardless of the input file type.

This requires the namelist to be Fortran readable, which the existing
implementation is not.  As a result, we need to break with the legacy
g-file format for this namelist.  This shouldn't be too much concern
as it isn't likely to be much use outside of EFIT.  Because of this,
restarts can only use g-files produced by EFIT-AI.

Some fixes were made to the restart code, including checks for a
consistent grid size and a change in the icinit numbers to be
consistent with exsisting logic.

Adding efitin and out1 namelists to the documentation as well.