Adding variables to m-files
Many of the chi squared values for measurments were missing.  Some
measurments and weights were also not present.  All the constraint
variables should now be included, aside from uncertainties (need to
be understood/fixed first).  Some reorganization and modernization was
done so that it is more easy to see what is or should be present.

Changing documentation to list the modules required to build or use
iris installations after a module purge (recommended).

Some additional clean-up and improvement of the code and documentation.