Bring certainty to uncertainties
Fixing the uncertainty calculation for flux loops and adding a
variable to save the uncertainty of the reference loop.

Renaming the uncertainties used by EFIT to better align with the
names of other measurements (for improved clarity).

Removing the nonlinear uncertainty calculations and associated
variables since this option is depreciated.

Updating documentation for uncertainty switches.

Some updates to configure scripts for Omega.