Protect more namelist reads and fix bug
More namelists will now error out when a problem occurs rather than
carrying on with eyes closed.  This caught a problem in namelists
created by OMFIT that contain an unsupported variable.  This should
be resolved in OMFIT (issue created) before merging to avoid disruption.

This also solves problems with the namelist reads that were raised by
PGI compilers (although not the MPI ones...)