Fix initializations from previous timeslice and limiter file
The legacy version uses the values in brsp from the previous timeslices,
potentially leading to inconsistent results depending on how timeslices
are distributed.  In this version we make the alternate approach of
not reusing past data the default, but allow the legacy behavior if
use_previous is set to true (can help some times converge).

The change in this version broke icinit<0 for snap modes. These icinit
options do not work for the legacy version in file mode, but they now
work for every mode in this version.  A new test for this has been added.

The limiter fix found in MR !34 has now been propagated back to the
shot range  used in tests.  This required updating most of the solutions,
which are now in perfect agreement with the legacy version, except for
cer fetching from snap (has minor differences).  Therefore we use the
legacy solutions as references again.

The debugging outputs in data_input.F90 have now been moved behind
the preprocessor debug flags to provide a cleaner terminal.  One
line that was unintentionally hidden for serial execution has been
brought back.  This motivated updates to all of the test outputs.

Additional options were added to prevent solutions from being
written that do not meet all of the error checks.  This can make it
easier to write a database of solutions that are high quality.  The
a-file alone can also be written out for use in a constraint knot
optimization loop.

A variety of code cleanup was performed, including the renaming
of chke.f90 to match the subroutine contained and move of an
unrelated subroutine to efit_utils.F90.