Updates and more specifics
Updating the m-file description to match the current outputs.

Including the input switches used to turn different output files
on/off.

Typo fix in in1 namelist.

The documentation for geqdsk and aeqdsk file contents probably need
to be checked in detail also, since the documentation seems pretty
old, but this is a great start!