Adding a new namelist and restructuring code
The "setup" namelist was added as an optional part of efit.input
files to specify configuration variables, such as the maximum
number of kfiles accepted and number of data points fetched from
ptdata.  This will hopefully make it possible to run efit on more
than 1001 timeslices at once and reconstruct longer shots accruately.
It also offers another way to set the location of tables to use, in
case users don't want to beholden to what is written in every k-file
(e.g. this supercedes the other options).  This is included in a test,
along with setting input variables with efit.input files (as opposed to
the commandline) closing an issue.

The code was restructured somewhat significantly to more clearly express
what some routines purpose is, to move closely related parts to the same
files, and eliminate duplicated code.  In particular, all of the
variable defaults and initializations have been moved to a single
subroutine set_defaults which separates out where variables get used
based on kdata, rather than being spread across 3 different routines and
many modules.

Also eliminated a bunch of unused variables and fixed documentation that
was no longer consistent with the code.

Fixed a bug in writing m-files and binary formatted g and a-files that
was introduced when the default for real variables was changed to 64 bit
in during compilation.