More documentation improvements
Adding files.rst to the website.

Doxygen calls in functions.rst and modules.rst are now being generated
by python and there should be less failures happening.

Adding the 2022 PPCF to the bibliography, fixing failures, and including
a blurb about this project on the main page of the website.

Most of the literal blocks have been used for lists so I am changing
the default to not use any code highlighting, but have added it to
scripts where code is shown.