Making single cyclic reduction the default solver
This method runs faster, produces identical results (at least as
as far as our regression tests show), and will be easier to make
parallel, so it seems to have no downsides!

Updated the corresponding documentation and all tests to use this
(default), except for one test (input) to pick up changes in the
double cyclic reduction as well.

Cleaned up some duplicate options in the IBUNMN implementation.