Setting the number of loops in single cyclic reduction based on grid size
Thanks to Oscar for catching this problem in larger test cases.

This makes the input option through dprobe.dat irrelevant.  Since it
could be problematic for changing grid sizes, it has been removed
from the files completely and is only kept as a dummy variable in
mhdin.dat for compatibility.