Move chkerr and allow limits to be set in experiment files
The chkerr subroutine is back in it's own file as requested by Scott.

The limits that used to be hard coded in this subroutine are based
on DIII-D.  To ensure the checks can be made to apply to any experiment,
the limits have been added to dprobe.dat and mhdin.dat files, with the
original values moved to defaults only.