Add a second require_plasma check
It turns out there are multiple ways that EFIT can end up producing
solutions without plasma.  The second way isn't as well understood yet,
but it doesn't hurt to have a second check.  It is broad enough to
catch all cases, but more efficient to avoid the plasma solve with the
first if possible.

Renaming inicur to set_init, since it does more than just setting the
initial current and in some cases it just leaves the current at zero
and sets other parameters instead.