Need a default for the device name
This will catch DIII-D dprobe.dat files that do not have a machinein
namelist included (older, not necessary since all defaults are defined
for DIII-D).

It will cause issues for other machines that do not have the namelist
or this variable included in it (older files need to be fixed).