Use error and chisq targets in solution checks instead of fixed values
This is a departure from how the solution checks were initially written
but better matches the error check message that gets printed out
(so there was inconsistency to start with).  It eliminates uncessary
extra variables, increases the applicability of checks, and aligns
with the stopping criteria in the code.  The default values have
also been changed to match the original checks.  There could be
some justification for using a looser checks than the convergence
targets (since that is also done for between shot EFITs), but it should
be better understood before it is implemented.

Fixing the printed messages for other solution checks to match the values used.

Also fixing and improving the documentation about the convergence criteria
and cleaning up some of the code.