Adding the final chi squared value to m-files
Currently the m-files have the chi squared values for every iteration,
but shapesurf updates the chi squared after the iterations finish.
Since both get used, it makes sense to have them in the m-files.

Fixed a bug in setting chi squared when exiting from the iconvr=2
checks that is also present in the original version of EFIT
because incorrect variables were used.

The tests are now able to compare the first line of the solution
summary that gets printed to the terminal (which has the final
chi squared).

Some code cleanup as well.