Bug fixes and new options
Bugs were uncovered when using different compilers and workflows
that differ from tests.  These include avoiding variable uses before
they are defined and ensuring namelist reads always start from the
beginning of a file in case some are missing or ordered differently.

Adding the option to change the maximum number of iterations that
can be set (mxiter) based on the size of allocated variables.  This
also requied a bug fix in write_m.

Setting up a new option (ierchk < 0) that skips running shapesurf
when fit hasn't converged, but still writes out an a-file.  This is
intended to speed up an optimization loop over constraint knot
positions.