Changing the validation check to be easier to use
Rather than having to change the check based on the types of
constraints, it now is simply on or off and the qualifiers get
applied based on other inputs (to differentiate mse and cer).
The parameter has been renamed from ivalid to req_valid.

Also updating the omega build to use a newly built system version
of d3lib.  This system is very unstable so more changes are expected.

Some further syntax cleanup.