Removing unneccesary subroutine calls
Oscar determined that the calls to green in currnt add unneccessary
calculations that slow down most cases.  This call is only needed
for fixed boundary calculations that use particular constraints so
it has been moved inside of a conditional that checks for such cases.
The sets2d call he also pointed out does not appear to be required for
any of the regression tests.  It has been commented out and the
conditions for which it is needed should be better understood before
it is added back.

Starting to add documentation for the ece namelist inputs.

Some additional code clean-up and documentation improvements.