Bug fixes for ece
This fixes another variable that needs to have global scope and
adds a check for when no ECE fitting is requested.

Also cleaning up some of the terminal output and removing explicit
reference to AI except for in the version label.

Some documentation improvements and code cleanup was also performed.