Fix and add tests for vessel current constraints
Several fixes were made to the vessel current constraints
(including an incorrect variable check in the original version)
and the implementation was cleaned up.  There may still be an
issue when MSE constraints are used with IFITVS=1, but these don't
appear to converge anyway.

2 tests were added for IVESEL=2 and IFITVS=1 (or IVESEL=3), but
they are not necessarily meaningful and could be replaced with a
better case.  A test for IVESEL=1 is also needed.

The documenation for modeling vessel currents has been completed
and stated more clearly.

Additional code clean-up and documentation improvement was performed.