Reformat mhdin.dat files and more documentation improvements
The mhdin.dat files now have all variables in the namelists and none
are appended.  This helps make the files more human readable (why
else use namelists?...).  The original versions are also retained for
posterity and because the revisions have not all been tested yet.

The IN3 variables that can also be read from outside of the namelist
have been described as such.  Hopefully this will fall out of favor,
but the documentation still covers legacy cases.

Some formatting in the documentation has been improved.