Small collection of code fixes
The most important fix is to reset the MSE channel weights to the
input values before writing k-files (when also solving for the
equilibrium).

Another fix is to the check for the existence of MSE before writing
k-files, which needs to be applied to each time separately.

A blank space that some compilers would print to the terminal caused
tests to fail.  That line has now been explicitily formatted to avoid
the problem.

The errctrl variables are now set as soon as the time array is defined
so that fewer error statements will reference an invalid time.  This
includes when writing k-files, which differs from the older code, so
the test outputs needed to be made consistent.

A handful of variables have been added to the documentation and many
other defintions have been improved.

So code formatting and reorganization was performed.