Adding a new option for outputing extra variables in g-files
The IPLCOUT option (introduced in 2004) adds extra variables to the end
of a g-file (before any namelists are included), but breaks the tradition
of keeping the files self descriptive (in terms of defining what variable
are present and their respective sizes).  Instead, this flag must be
read in from the snap or k-file and the f-coil and e-coil sizes must
be read from dprobe.dat before the g-file can be fully described.  These
challenges are now described in the source code, namelist, and output
file documentation.

In order to relax the criteria for interpreting the extra variables
without breaking the rest of the format, a new option has been added
which only adds the 2D plasma current.

Several other documentation and syntax fixes were made, including more
accurate code for reading the output files (now in modern Fortran).