Pointing to the location where the master set of snapfiles lives
This is a sub-directory within the support files where all named
snapfiles are collected (at least for DIII-D).  EFIT checks this
area if it does not find the requested snap extension in the
current working directory.  Documentation and comments have been
added to better explain this behavior.

Some additional documentation improvement and clean-up was also
performed.