Making a super namelist to help with database creation
The IN0 namelist can now be specified in an efit.input file to
override parameters in the IN1 namelists.  This avoids having
to edit thousands of k-files for thousands of shots when they
may have already been generated without the correct parameters.
This only includes 3 variables at this time, but could be expaneded
to include all of the scalar control parameters if they are needed.

Also adding more recognizable errors for when a required input
file is not found.