Modifying ierrchk options and adding iout to the in0 namelist
ierchk option 2 (which was redundant) has been changed to match
option 3, with the exception of leaving out the error and chi
squared checks.  This is intended to be used with reconstructions
that are already trusted (like kinetic EFITs) in spite of failing
to converge by this definition.  In such cases the other checks
can still be valuable for catching problems.  chkerr was also
substantially cleaned up.

iout options are now easier to control through efit.input files.
The in0 namelist is now included in documentation as well (was
missed in previous commits).