Improvements primarily for documentation purposes
This includes fixes to the docs, in addition to some code commenting
and clean up.

The validity parameter bounds have also been added to DIII-D and NSTX
experiment files.  For DIII-D these match the default values and are
only added for completeness/accessibility.  For NSTX the values were
determined by Alexei to be appropriate for the database generation.
These values have not yet been propogated to the support files on
NERSC, GA, PPPL, or other clusters so they will only be available if
personal copies of the Green function tables are made.