Bug fixes and changes to the MSE libraries
A but with setting zelip that impacted experiments other than DIII-D
has been corrected.

The variable size allocated for the response matrix solve (hardcoded)
has been increased to allow for the large set of diagnostics on ITER.

The configure files have been updated to point to new versions of the
MSE libraries on Iris and Omega.  This changed how the edges of beam
blips are handled, which causes small changes to the results (so the
tests have been updated).  The MSE devs also recommended changing to
set msebkp=7, which will also be propagated to between shot EFITs soon.
For now only the snap tests include this change.  It may be worth
updating the k-files in other tests for consistency, but it isn't
essential.

Some code modernization and documentation improvements have also
been made.