Remove non-standard Fortran functions to improve GNU compiler support and optimization
EFIT used a handful of math functions that were not always part of the Fortran
standard (cosd, sind, tand, atand, and atan2d).  This required a special
flag to be passed to the GNU compilers (-fdec-math) which is not supported by
older compilers (GCC<7) and also appears to have been a primary factor behind
why some of the optimization in -Ofast was not working (could have been other
code improvements over the past year as well).

A configuration file for the default GCC version on Iris (4.7.2) has been
added as a result.

Some additional code, documentation, and test result clean up was performed
as well.