Fix HDF5 linking on NERSC
Cray likes to set up their libraries differently so modules need to
be used for everything to work properly.  This is now setup to work
for HDF5.  NetCDF should probably be handled similarly, but it looks
like some modules are still missing and handling that without modules
is still working (for now).