Cleaning up the spline routines and renaming a-file variables
Duplicate spline routines were removed and the remaining ones have
been modernized.

Variables that are output in a-files now have the same name, or as
close as possible, to what is displayed in efitviewer.  These names
are more intuitive and should be more relatable to users now.

Some additional clean-up and documentation improvments are also
included.