Avoid repeated initialization of cyclic reduction and correct error metric with solution relaxation
Oscar found that the initialization step for the cyclic reduction algorithm was being called
every iteration even though it wasn't needed after the first.  Removing this increases the
performace of pflux by approximately 20%.

The RELAX option allows EFIT to reduce the variation produced with each solver iteration by
combining it with the prior result.  Since the ERROR metric compares the maximum change in the
poloidal flux between the current and previous iteration, a small RELAX can artificially appear
to converge by reducing the change in psi.  To account for this, the reported error is now
divided by RELAX.  This required updating the test results for any case with RELAX<1.0.

Cases without Li beam won't have these outputs written to m-files any longer, (part of the MSE
outputs) to reduce the amount of empty data being produced.

Some code clean-up was done in the basis function routines and related documentation was also
improved.