Update documentation and NetCDF labels to describe F-coil measurements in Amp-turns
The F-coil current pulled from PTDATA is multiplied by the number of
turns (see line 565 of measurements.F90) and used in this form throughout
the rest of the code.  As a result, the documentation and labels have
been mis-characterized.

This was first recognized by Joey in a DIII-D CWG issue.