Initial setup of file writes to OMAS HDF5
All quantities have been checked for accuracy against the OMFIT
conversion when possible.  Some fields have not yet been added
and there are some minor differences (unused ods are not written,
boundary_separatrix is used instead of boundary, and others are
anticipating IMAS 3.39 which hasn't been released yet)

This has not been added to the tests yet, but will be soon.
Afterwards the tests can be made more rigorous with h5diff.

The implementation only works with serial execution at the moment
with no warning about issues with MPI.  This will be added with
parallel HDF5 support.