Need to setup OMAS output file with the first write
The first time slice could fail to make it to the write statement
so that is not a robust choice.  Instead it is now the first to
enter the subroutine, regardless of the time.  This complicates an
MPI implementation, so the best solution is to silence warnings for
that data already existing if it has been written by a different
processor.

The NSTX and ITER results also changed now that they aren't using
DIII-D spatial dimensions to define inboard, outboard, top and
bottom parts of the limiter.  The test results have been updated
accordingly.

Now that Cori has retired and we are stuck with the wild west that
is Perlmutter, the definitions are no longer needed in configure
scripts.  RIP Cori.