Setup tests to use HDF5 files produced by EFIT
This replaces the need to run the OMAS conversion through OMFIT for
any new test case (that is only needed for cases that depend on a
previous solution now).  The tests do not use h5diff yet, but when
that is added, the reference HDF5 solution should be updated.

Most of the HDF5 tests have been disabled for parallel builds, since
only the serial HDF5 writes have been setup, so far.

A warning message is now printed if OMAS output is requested in a
parallel run.

More provenance info has been added to the OMAS write, including
the version of EFIT that was run and the creation date.

More of the input namelist variables are saved so that when they
are re-written to an OMAS file the same case can be re-run.

The limiter side definitions were added to the mhdin.dat namelists
that are read in efund.

The input (k-file) creation from snapfiles modes (5 and 6) were
fixed so that they can be run in parallel (the only benefit is to
write the files in parallel and it is not clear that this justifies
the MPI overhead, but the option is available now).

Documentation focused on adding new tests has been updated.

A few other minor code fixes were made.