Tweak the boundary algorithm to be less sensitive to FP changes
Small FP differences between the original EFIT and EFIT-AI broke
the boundary finder and caused EFIT-AI not to converge.  To protect
against this, additional checks were added to the exit criteria
which should nudge points outside, but very close to the separatrix
so that they will be inside for the next iteration.