Remove invalid transpose for 2D data and improve test robustness
2D data should be stored in OMAS with [R,Z] orientation, but a bug in
OMFIT produced the expecation that this was not the case.

Tests were previously sensitive to whether CMAKE variables were defined
by ON/OFF or TRUE/FALSE (both valid) and their case.  All options
allowed by CMAKE should now work properly in the tests.

The changes to configure NETCDF on hip were reverted, as the issue has
not appeared on other branches (still not understood).