Bug fixes
Fixing several compile and optimized runtime bugs that were caught
by the PGI compiler.

Also fixing the half baked attempt to avoid writing so many zeros
for the MSE data.  Found a bug in a NetCDF defined variable name
that had been around for a while as well (only impacted the title
of files).