Rename variables instead of duplicating
Since all of the m-file outputs that had different names were new
additions with EFIT-AI, there should not be very much impact to
changing the names (OMFIT has been updated).  This improves agreement
with rt-EFIT as well, since that seems to be where the MDS+ names
came from (in 2018).

The internal variable names were also updated if this was the only
place they were input or output, but in 2 cases the variables were
part of IN1 so the name change only happens when writing to m-files.