Initial setup for extending the j_tor constraint
The fit values of this constraint are now computed and included in the
m-file output.  This has been checked in cases without toroidal
rotation and looks accurate.

The weight used for these constraints (constant) is also output in
the m-files now.

Place holders have been made for outputing the uncertainty and chi
squared for the j_tor fits as well.  Options to use these still need
to be added.

An alternative form of the constraint may be added as part of ongoing
work as well.

Code fixes, cleanup, and documentation improvements have been made as
part of this commit.  Notably, there appeared to be an indexing error
in this constraint for cases with flow that has now been fixed.