Add options to set weight (per point) and uncertainty for j_tor constraint
The alternate inputs are only used if they are set (default behavior
is the same).  The chi squared metric for fitting j_tor is computed
if the uncertainty is input (expecting the same units).

Some cleanup and documentation improvement is also included.

This completes the work currently planned for this constraint.