Initial commit for knot optimization improvements
This adds a new option similiar to the existing auto_knot that will return
as soon as convergence criteria are met, instead of optimizing further.

Some new options and coding improvements have been made that also apply to
auto_knot, including providing a limit to the range over which knots are varied
and setting a default tolerance that should work better when the number of knots
and the ranges varies.  The knot search range currently applies to all points, but
will be generalized so that some knots can vary more than others.

Some other code fixes are included here as well.