Adding a new knot optimization method with random perturbations
This is currently setup as new option with KAUTOKNT=3, but once
more testing is complete, it will probably be worth choosing between
this and knot_opt before merging into main.

Tests with shot 174082 show signs that this method could be fast
and effective for optimizing the knots.  I tried running 2 timeslices
10 times and on average it only took around 3 attempts for this method
to converge.  This seems to be more efficient than optimizing one knot
at a time, but more cases should be tested.

It also has the added benefit of being much simpler so an MPI
implementation should be straight forward.