Several code fixes, clean-up, and variable renaming
The most noticable change here will be some different variables
printed to the terminal when EFIT finishes.  Some of the names
were changed to be more descriptive or better match the variables
being output.  In a few cases these name changes were propagated
through the rest of the code as well.  Some of the variables were
replaced with similar, but more useful quantities, such as the
center of the current and q* for the magnetic axis and q95.  Lastly,
a couple of variables that did not seem useful were replaced with
the GS residual and total chi squared value (including diamagntic,
pressure, and current constraints).

The constant minimum flux loop uncertainty (sicont) was put inside
a conditional so that it is only applied when errsil=0.  This allows
for comparison with older versions of EFIT.

The OMAS test input files for restart cases were replaced with
versions that have the correct COCOs for the sign of psi
(courtesy of a PR in OMFIT).  So the temporary hack in data_input
which allowed tests with the wrong sign to pass has been removed.

Input and output variable documentation was also updated.