Update ITER machine files
The 1 directory now contains a description where only the CS and
plasma shaping coils are f-coils and all other (VS and TF busbar)
are e-coils.  This is intended to help basic modeling when coil
currents are not known.  Directory 2 has a more realistic description
with CS coils as e-coils, shaping and VS coils as f-coils, and the
TF busbar as an a-coil.

The coil packs in the middle of the CS are now represented as a single
coil in EFIT, which should better resemble how they are connected on
ITER.