Update DIII-D limiter and ITER coil and flux loop descriptions
The DIII-D limiter now includes the geometry for the upcoming campaign.

A new ITER description was created from the latest MD available.  This
also takes into account that the flux loop types are mis-labeled.  Some
of these may not be accurate if they have errors in ITER's MD (as the
mis-labeling would suggest), but it's the best currently available.