Fixes for fetching density signals from MDS+
The MDS+ tree for interferometers was recently reorganized so that
relative paths can no longer be used.  This is how amdata was setup
in EFIT so we had to switch to using the absolute path.  These densities
are the only values EFIT currently fetches from MDS+, but the change was
isolated to these signals in case the behavior is unwanted with a
future extension of the MDS+ usage.  These changes have already been
made in the version of EFIT running on Iris and Omega (to avoid down-
time from it's use on DIII-D).

The fitweight.dat file was also updated to match the latest changes on
DIII-D and one of the mhdin.dat files was re-formatted.